chore: Fix issue with admin account used
This commit fixes an issue in the accessory list component where the admin account was not being properly handled. It updates the `handleAdminAccountUsed` method to be asynchronous and adds a call to `resetConfig` in order to reset the plugin's configuration. This ensures that the user will use guest admin account instead.